+++ src/nalu/annex_b_nalu.hpp
@@ -9,17 +9,42 @@
 #include <cstdint>
 #include <ostream>
 
-#include "nalu_type_from_header.hpp"
-#include "nalu_type_to_string.hpp"
-#include "nalu_type.hpp"
+#include "type_from_header.hpp"
+#include "type_to_string.hpp"
+#include "type.hpp"
 
 namespace nalu
 {
 /// Small struct representing an Annex B NALU
 struct annex_b_nalu
 {
-    /// @return True if the NALU is valid otherwise false
+    annex_b_nalu() :
+        m_data(nullptr),
+        m_size(0),
+        m_start_code_size(0),
+        m_type(nalu::type::unspecified0)
+    {
+        assert(!is_valid());
+    }
+
+    annex_b_nalu(const uint8_t* data, uint32_t size, uint32_t start_code_size,
+                 nalu::type type) :
+        m_data(data),
+        m_size(size),
+        m_start_code_size(start_code_size),
+        m_type(type)
+    {
+        assert(m_data != nullptr);
+        assert(m_size >= m_start_code_size);
+        assert(m_start_code_size == 3U || m_start_code_size == 4U);
+    }
+
     operator bool() const
+    {
+        return is_valid();
+    }
+
+    bool is_valid() const
     {
         if (m_data == nullptr)
             return false;
@@ -28,34 +53,33 @@ struct annex_b_nalu
             return false;
 
         // A valid start code is either 3 or 4 bytes
-        if (m_startcode_size != 3 && m_startcode_size != 4)
+        if (m_start_code_size != 3 && m_start_code_size != 4)
             return false;
 
         return true;
     }
 
     /// Pointer to the NALU data
-    const uint8_t* m_data = nullptr;
+    const uint8_t* m_data;
 
     /// Size of NALU in bytes
-    uint32_t m_size = 0;
+    const uint32_t m_size;
 
     /// Size of the Annex B start code
-    uint32_t m_startcode_size = 0;
+    const uint32_t m_start_code_size = 0;
 
     /// The type of the NALU
-    nalu_type m_type;
+    const nalu::type m_type;
 };
 
 /// Output operator for the annex_b_nalu struct
 inline std::ostream& operator<<(std::ostream& os, const annex_b_nalu& nalu)
 {
-    assert(nalu);
-
-    os << "nalu::annex_b_nalu: m_data = " << (void*) nalu.m_data << " "
-       << "m_size = " << nalu.m_size << " m_startcode_size = "
-       << nalu.m_startcode_size << " (type = "
-       << nalu_type_to_string(nalu.m_type) << ")";
+    os << "annex_b_nalu:" << std::endl
+       << " data = " << (void*) nalu.m_data << std::endl
+       << " size = " << nalu.m_size << std::endl
+       << " start_code_size = " << nalu.m_start_code_size << std::endl
+       << " type = " << type_to_string(nalu.m_type) << std::endl;
 
     return os;
 }