+++ src/nalu/error_tags.hpp
@@ -5,12 +5,12 @@
 // The copyright notice above does not evidence any
 // actual or intended publication of such source code.
 
-nalu_ERROR_TAG(unused,
+NALU_ERROR_TAG(unused,
                "Reserved! Errors should never be zero :)")
 
-nalu_ERROR_TAG(no_annex_b_nalu_data_found,
+NALU_ERROR_TAG(no_annex_b_nalu_data_found,
                "No Annex B NALU data found")
 
-nalu_ERROR_TAG(garbage_found_in_nalu_data,
+NALU_ERROR_TAG(garbage_found_in_nalu_data,
                "Garbage data was found in the NALU data, the Annex B "
                "NALU must start with a 3 or 4 byte startcode")