+++ src/nalu/type_to_string.hpp
@@ -7,23 +7,23 @@
 
 #pragma once
 
-#include "nalu_type.hpp"
+#include "type.hpp"
 
+#include <cassert>
 #include <string>
 
 namespace nalu
 {
 /// @return Human readable message corresponding to an error code
-inline std::string nalu_type_to_string(nalu_type type)
+inline std::string type_to_string(nalu::type type)
 {
     switch (type)
     {
 #define NALU_TYPE_TAG(value,id,msg,vcl)      \
-                case nalu_type::id: return std::string(msg);
-#include "nalu_type_tags.hpp"
+                case nalu::type::id: return std::string(msg);
+#include "type_tags.hpp"
 #undef NALU_TYPE_TAG
     }
-
     assert(0 && "Invalid NALU type received!");
     return "";
 }