+++ NEWS.rst
@@ -6,7 +6,17 @@ every change, see the Git log.
 
 Latest
 ------
-* tbd
+* Major: Renamed ``nalu_type_from_header.hpp`` -> ``type_from_header.hpp``.
+* Major: Renamed ``nalu_type_from_header`` -> ``type_from_header``.
+* Major: Renamed ``nalu_type_to_string.hpp`` -> ``type_to_string.hpp``.
+* Major: Renamed ``nalu_type_to_string`` -> ``type_to_string``.
+* Major: Renamed ``nalu_type.hpp`` -> ``type.hpp``.
+* Major: Renamed ``nalu_type`` -> ``type``.
+* Major: ``annex_b_nalu`` now has a two constructors, one for constructing valid
+  nalus and one for constructing invalid nalus. The validity of a nalu can be
+  checked with the ``is_valid`` member function and the ``operator bool()``.
+* Major: Renamed ``annex_b_nalu::startcode_size`` ->
+  ``annex_b_nalu::start_code_size``.
 
 3.0.0
 -----